Seperate Rqes parameters
